Woz to TypeScript
